moved CSV export button to its own component and into MainView
